requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.4.0
Pillow==10.0.0
scikit_learn==1.3.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.29.1
